+++ ChangeLog.md
@@ -1,3 +1,9 @@
+## Changes between ifm3d-ros 0.2.0 and 0.3.0
+
+* Added `Dump` Service
+* Added `Config` Service
+* Added `Trigger` Service
+
 ## Changes between ifm3d-ros 0.1.0 and 0.2.0
 
 * Updates to CMakeLists.txt to support Ubuntu 14.04 and ROS Indigo